Implemented FenStringModal
